+++ apps/client/src/app/pages/home/assets/facebook-logo.png
