+++ apps/client/src/app/pages/home/assets/instagram-logo.png
